requirements.txt
discord.py
pyyaml
polib
youtube-dl
ffmpeg-python
pynacl
requests